requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
google-api-core==1.3.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
httplib2==0.11.3
idna==2.7
numpy==1.15.0
oauth2client==4.1.2
pandas==0.23.3
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==3.4.2
ruamel.yaml==0.15.47
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
